feat: use wallet daemon client with webrtc rpc transport
