+++ example/src/App.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import { Typography, Container, Button, Paper, Stack, Divider, TextField } from '@mui/material';
-import { TransactionSubmitRequest } from '../../src/providers';
+import { SubmitTransactionRequest } from '../../src/providers';
 import Header from './Header';
 
 export default function App() {
