+++ src/providers/wallet_daemon/index.ts
@@ -1,8 +1,5 @@
 export { TariConnection } from "./webrtc";
 
-export default function test() {
-  console.log("test")
-}
 
 export * from './tari_permissions'
 export * from './provider'
