+++ src/providers/wallet_daemon/provider.ts
@@ -1,8 +1,15 @@
 import {TariPermissions} from "./tari_permissions";
 import {TariConnection} from "./webrtc";
 import {TariProvider} from '../index';
-import {TransactionSubmitRequest, TransactionResult, TransactionStatus, TransactionSubmitResponse} from '../types';
+import {SubmitTransactionRequest, TransactionResult, TransactionStatus, SubmitTransactionResponse} from '../types';
 import {Account} from "../types";
+import {
+    WalletDaemonClient,
+    stringToSubstateId,
+    Instruction,
+    TransactionSubmitRequest, SubstateType, SubstatesListRequest
+} from "@tariproject/wallet_jrpc_client";
+import {WebRtcRpcTransport} from "./webrtc_transport";
 
 export const WalletDaemonNotConnected = 'WALLET_DAEMON_NOT_CONNECTED';
 export const Unsupported = 'UNSUPPORTED';
@@ -19,11 +26,11 @@ export type WalletDaemonParameters = {
 export class WalletDaemonTariProvider implements TariProvider {
     public providerName = "WalletDaemon";
     params: WalletDaemonParameters;
-    connection: TariConnection;
+    client: WalletDaemonClient;
 
-    private constructor(params: WalletDaemonParameters, connection: TariConnection) {
+    private constructor(params: WalletDaemonParameters, connection: WalletDaemonClient) {
         this.params = params;
-        this.connection = connection;
+        this.client = connection;
     }
 
     static async build(params: WalletDaemonParameters): Promise<WalletDaemonTariProvider> {
@@ -32,49 +39,54 @@ export class WalletDaemonTariProvider implements TariProvider {
         allPermissions.addPermissions(params.optionalPermissions);
         console.log({allPermissions});
         let connection = new TariConnection(params.signalingServerUrl, params.webRtcConfig);
-        await connection.init(allPermissions, params.onConnection);
-        return new WalletDaemonTariProvider(params, connection);
+        const client = WalletDaemonClient.new(WebRtcRpcTransport.new(connection));
+        await connection.init(allPermissions, (conn) => {
+            params.onConnection?.();
+            if (conn.token) {
+                client.setToken(conn.token);
+            }
+        });
+        return new WalletDaemonTariProvider(params, client);
     }
 
     public get token(): string | undefined {
-        return this.connection.token;
+        return (this.client.getTransport() as WebRtcRpcTransport).token();
     }
 
     public get tokenUrl(): string | undefined {
-        if (this.connection.token) {
-            const name = this.params.name && encodeURIComponent(this.params.name) || '';
-            const token = this.connection.token;
-            const permissions = JSON.stringify(this.params.permissions);
-            const optionalPermissions = JSON.stringify(this.params.optionalPermissions);
-
-            return `tari://${name}/${token}/${permissions}/${optionalPermissions}`
+        if (!this.token) {
+            return undefined;
         }
-        return undefined;
+
+        const name = this.params.name && encodeURIComponent(this.params.name) || '';
+        const token = this.token;
+        const permissions = JSON.stringify(this.params.permissions);
+        const optionalPermissions = JSON.stringify(this.params.optionalPermissions);
+
+        return `tari://${name}/${token}/${permissions}/${optionalPermissions}`
     }
 
     public isConnected(): boolean {
-        return this.connection.isConnected();
+        return (this.client.getTransport() as WebRtcRpcTransport).isConnected();
     }
 
     public async createFreeTestCoins(): Promise<Account> {
-        const method = "accounts.create_free_test_coins";
-        const res = await this.connection.sendMessage(method, this.connection.token, {
-            account: null,
+        const res = await this.client.createFreeTestCoins({
+            account: {Name: "template_web"},
             amount: 1000000,
             max_fee: null,
             key_id: 0
-        }) as any;
+        });
         return {
             account_id: res.account.key_index,
-            address: res.account.address.Component,
+            address: (res.account.address as { Component: string }).Component,
             public_key: res.public_key,
             resources: []
         };
     }
 
     public async getAccount(): Promise<Account> {
-        const method = "accounts.get_default";
-        const {account, public_key} = await this.connection.sendMessage(method, this.connection.token, {}) as any;
+        const {account, public_key} = await this.client.accountsGetDefault({}) as any;
 
         return {
             account_id: account.key_index,
@@ -86,38 +98,37 @@ export class WalletDaemonTariProvider implements TariProvider {
     }
 
     public async getAccountBalances(componentAddress: string): Promise<unknown> {
-        const method = "accounts.get_balances";
-        const args = {ComponentAddress: componentAddress};
-        const res = await this.connection.sendMessage(method, this.connection.token, args);
-
-        return res;
+        return await this.client.accountsGetBalances({account: {ComponentAddress: componentAddress}, refresh: true});
     }
 
-    public async getSubstate(_substate_address: string): Promise<unknown> {
-        // TODO: the wallet daemon should expose a JRPC method to retrieve any substate
-        throw Unsupported;
+    public async getSubstate(substate_id: string): Promise<unknown> {
+        const substateId = stringToSubstateId(substate_id);
+        return await this.client.substatesGet({substate_id: substateId});
     }
 
-    public async submitTransaction(req: TransactionSubmitRequest): Promise<TransactionSubmitResponse> {
+    public async submitTransaction(req: SubmitTransactionRequest): Promise<SubmitTransactionResponse> {
         const params = {
             signing_key_index: req.account_id,
-            fee_instructions: req.fee_instructions,
-            instructions: req.instructions,
-            inputs: req.required_substates,
+            fee_instructions: req.fee_instructions as Instruction[],
+            instructions: req.instructions as Instruction[],
+            inputs: req.required_substates.map((s) => ({
+                // TODO: Hmm The bindings want a SubstateId object, but the wallet only wants a string. Any is used to skip type checking here
+                substate_id: s.substate_id as any,
+                version: s.version
+            })),
             override_inputs: false,
             is_dry_run: req.is_dry_run,
             proof_ids: [],
             min_epoch: null,
             max_epoch: null,
-        };
-        const res = await this.connection.sendMessage<any>("transactions.submit", this.connection.token, params, 10);
+        } as TransactionSubmitRequest;
+        const res = await this.client.submitTransaction(params);
 
         return {transaction_id: res.transaction_id};
     }
 
     public async getTransactionResult(transactionId: string): Promise<TransactionResult> {
-        const params = {transaction_id: transactionId};
-        const res = await this.connection.sendMessage("transactions.get_result", this.connection.token, params) as any;
+        const res = await this.client.getTransactionResult({transaction_id: transactionId});
 
         return {
             transaction_id: transactionId,
@@ -127,8 +138,18 @@ export class WalletDaemonTariProvider implements TariProvider {
     }
 
     public async getTemplateDefinition(template_address: string): Promise<unknown> {
-        const params = {template_address};
-        return await this.connection.sendMessage("templates.get", this.connection.token, params);
+        return await this.client.templatesGet({template_address});
+    }
+
+    public async listSubstates(
+        template: string | null,
+        substateType: SubstateType | null
+    ) {
+        const resp = await this.client.substatesList({
+            filter_by_template: template,
+            filter_by_type: substateType
+        } as SubstatesListRequest);
+        return resp.substates as any[];
     }
 }
